Do an initial pass on refactoring out ElementContext from WindowContext
